chore: repository 코드 정리
