+++ Boolti/Boolti/Sources/Network/Repositories/Auth/AuthRepository.swift
@@ -7,13 +7,21 @@
 
 import Foundation
 
-import KakaoSDKAuth
-import KakaoSDKUser
-import SwiftJWT
-import RxKakaoSDKUser
 import RxSwift
-import RxMoya
-import Kingfisher
+import KakaoSDKUser
+import RxKakaoSDKUser
+import SwiftJWT
+
+protocol AuthRepositoryType {
+
+    var networkService: NetworkProviderType { get }
+    func fetchTokens() -> (String, String)
+    func fetch(withProviderToken providerToken: String, provider: OAuthProvider) -> Single<SignupConditionEntity>
+    func signUp(provider: OAuthProvider, identityToken: String?) -> Single<Void>
+    func logout() -> Single<Void>
+    func userInfo() -> Single<Void>
+    func resign(reason: String, appleIdAuthorizationCode: String?) -> Single<Void>
+}
 
 final class AuthRepository: AuthRepositoryType {
