+++ Boolti/Boolti/Sources/Network/Repositories/Auth/RefreshAuthRepository.swift
@@ -9,7 +9,6 @@ import Foundation
 
 import Moya
 import RxSwift
-import RxMoya
 
 final class RefreshAuthRepository {
