--- Boolti/Boolti/Sources/Network/Repositories/Protocols/QRRepositoryType.swift
@@ -1,15 +1,0 @@
-//
-//  QRRepositoryType.swift
-//  Boolti
-//
-//  Created by Juhyeon Byun on 2/14/24.
-//
-
-import RxSwift
-
-protocol QRRepositoryType {
-    
-    var networkService: NetworkProviderType { get }
-    func scannerList() -> Single<[QRScannerEntity]>
-    func qrScan(concertId: Int, entranceCode: String) -> Single<Bool>
-}
